+++ src/routes/irbs/PrincetonCovidDisinformationIRB.svelte
@@ -1,0 +1,208 @@
+<style>
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+.research-is-voluntary, h1, h2, p, dl {
+    color: var(--irb-text-color, var(--color-marketing-gray-70));
+}
+
+.research-is-voluntary {
+    font-size: 16px;
+    margin-bottom: 28px;
+}
+
+h1 {
+    font-size: 24px;
+    font-family: "Zilla Slab", -apple-system, BlinkMacSystemFont, "Segoe UI",
+      Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji",
+      "Segoe UI Symbol";
+    margin-bottom: 28px;
+}
+h2 {
+    font-size: 16px;
+    font-weight: 700;
+    line-height: 24px;
+    font-family: "Inter", -apple-system, BlinkMacSystemFont, "Segoe UI",
+      Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji",
+      "Segoe UI Symbol";
+}
+
+p {
+    font-size: 14px;
+    line-height: 21px;
+    margin-bottom: 20px;
+}
+.mzp-u-list-styled {
+    font-size: 14px;
+    line-height: 21px;
+}
+
+dl {
+    margin-bottom: 32px;
+}
+
+dt {
+    font-weight: 900;
+    text-transform: uppercase;
+}
+
+dd {
+    margin-bottom: 1rem;
+}
+
+</style>
+
+<h1>Rally Study Privacy Consent Notice</h1>
+
+<div class='research-is-voluntary'>
+    <div>Your informed consent is being sought for research.</div>
+    <div>
+        Participation in the research is voluntary.
+    </div>
+</div>
+
+<dl>
+    <dt>Academic Institution</dt>
+    <dd>Princeton University</dd>
+    <dt>Study Name</dt>
+    <dd>COVID-19 and Political Information Flows</dd>
+    <dt>Principal Investigator</dt>
+    <dd>
+      Jonathan Mayer, Assistant Professor of Computer Science and Public
+      Affairs
+    </dd>
+  </dl>
+
+<h2>IRB Approval</h2>
+<p>
+    This study has been approved by the Institutional Review Board for Human Subjects at Princeton University.
+</p>
+
+<h2>Study Purpose</h2>
+<p>
+    The study assesses how individuals are exposed to, consuming, sharing, and acting on health and political 
+    information and misinformation during the COVID-19 pandemic.
+</p>
+
+<h2>Study Benefits to Society</h2>
+<p>
+    This study will help identify information distribution pathways for truthful and inaccurate information during the 
+    COVID-19 pandemic. Understanding these pathways is crucial to designing public health interventions and addressing online misinformation.
+</p>
+
+<h2>Study Procedures</h2>
+<p>
+    During the study, your browser will automatically send data to Mozilla about how you engage with health and news websites, 
+    in addition to standard Firefox telemetry. We have taken a number of steps to minimize and protect this data, which we describe below. 
+    You may also be asked to participate in periodic surveys. 
+    If you choose to provide demographic information for Mozilla Rally research purposes, 
+    that information will also be used for this study.
+</p>
+
+<p>
+    Your participation in the study will conclude when data collection is completed or, 
+    if you choose to discontinue participation, when you make that choice. 
+    Data collection will last no longer than 6 months, 
+    and we will not retain data that is linkable to your browser for longer than 6 months.
+</p>
+
+<h2>Leaving the Study</h2>
+<p>
+    You can leave the study at any time from the Mozilla Rally options page. 
+    To access the page, click on the Rally button <img style="padding: 0 4px;" width="18" src="img/rally-toolbar-icon.svg" alt="rally icon" /> in your browser toolbar. 
+    The button is usually near the top right of the browser window. 
+    If you have removed the Rally button from your toolbar, 
+    you can also access the Rally options page from the browser’s Add-ons settings. 
+    The Rally options page will show you a list of studies that you are currently enrolled in. 
+    You can leave this study by clicking the “Leave Study” button on the associated study card.  
+    If you leave the study, the study’s browser extension will be automatically uninstalled, 
+    removing the study code and data from your browser. 
+    We will also automatically delete the data that your browser has submitted for the study, 
+    unless you allow us to retain that data.
+    You may not be able to rejoin the study if you leave. 
+    If you leave Rally, you will also leave the studies that you are enrolled in, including this study.
+</p>
+
+<h2>Study Risks and Alternative Procedures</h2>
+<p>
+    There is minimal risk associated with participating in the study. 
+    There are no alternative means of participating in the study that may be advantageous to you. 
+</p>
+
+<h2>Privacy</h2>
+<p>
+    We will protect the information that you provide in several ways: minimizing the data we collect, 
+    keeping the data we collect secure, and limiting the information we use and disclose.
+</p>
+
+<p>
+    <em>Data minimization:</em> We will limit our data collection to information needed to carry out our study. 
+    We will not collect your name, email address, or similar contact information. 
+    We also will make no effort to identify you on the basis of information you provide. 
+    When collecting data about your web browsing, we will use coarse and aggregate measures as much as possible. 
+    For instance, when collecting data about your visit to a webpage, 
+    we will only collect a <em>time window</em> in which the event occurred instead of the exact time, 
+    and we will only collect the <em>domain name</em> and <em>page type</em> for the webpage rather than the full URL. 
+    We will also only collect activity specific to a website 
+    if we have previously categorized the website as health or news related.
+</p>
+
+<p>
+    <em>Data security:</em> We will encrypt the data that you provide both in transit and at rest. 
+    We will only allow access to Princeton researchers and Mozilla personnel who have specific responsibilities in carrying out the study.
+</p>
+
+<p>
+    <em>Data use and disclosure:</em> We will only use the data you provide to carry out this academic study. 
+    We will not disclose information that identifies you or that could reasonably be used to identify you.
+</p>
+
+<h2>Compensation</h2>
+<p>There is no compensation for participation in this study.</p>
+
+<h2>Who to Contact with Questions</h2>
+<p>
+    Please contact the Principal Investigator, Jonathan Mayer (<a
+    href="mailto:covid-19-study@lists.cs.princeton.edu"
+    >covid-19-study@lists.cs.princeton.edu</a
+    >), with any questions about the study.
+</p>
+
+<p>
+    If you have questions regarding your rights as a research subject, or
+    if problems arise which you do not feel you can discuss with the
+    Investigator, please contact the Institutional Review Board for Human
+    Subjects at Princeton University at:
+</p>
+
+<p>
+    Assistant Director, Research Integrity and Assurance<br />
+    (609) 258-8543<br />
+    <a href="mailto:irb@princeton.edu">irb@princeton.edu</a>
+</p>
+
+<h2>Summary of Consent</h2>
+
+<p>I understand the information that was presented and that:</p>
+<ul class="mzp-u-list-styled">
+    <li>My participation is voluntary.</li>
+    <li>
+    Refusal to participate will involve no penalty or loss of benefits
+    to which I am otherwise entitled.
+    </li>
+    <li>
+    I may discontinue participation at any time without penalty or loss
+    of benefits.
+    </li>
+    <li>
+    I do not waive any legal rights or release Princeton University or
+    its employees from liability for negligence.
+    </li>
+    <li>
+    I am over the age of 18 and located in the United States.
+    </li>
+</ul>
+<p>I hereby give my consent to be the subject of the research.</p>
+
+